backend/requirements.txt
# ============================================================
# Core Framework
# ============================================================
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
websockets>=12.0
pydantic-settings>=2.1.0
pydantic>=2.0.0

# ============================================================
# LLM — Local (Ollama) + Cloud (OpenAI / Bedrock)
# ============================================================
ollama>=0.4.0
langchain>=0.3.0
langchain-ollama>=0.2.0
langchain-community>=0.3.0
langchain-chroma>=0.1.0
langchain-openai>=0.2.0        # optional OpenAI fallback

# ============================================================
# RAG & Embeddings
# ============================================================
chromadb>=0.5.0
sentence-transformers>=3.0.0
pypdf>=4.0.0
python-docx>=1.1.0

# ============================================================
# Voice
# ============================================================
openai-whisper>=20231117
sounddevice>=0.4.6
soundfile>=0.12.1
numpy>=1.24.0
scipy>=1.11.0
pydub>=0.25.1
pyttsx3>=2.90
gTTS>=2.5.0

# ============================================================
# Database
# ============================================================
sqlalchemy>=2.0.25
aiosqlite>=0.19.0

# ============================================================
# Security — new in v2.0
# ============================================================
PyJWT>=2.8.0                   # JWT tokens (auth.py)
cryptography>=42.0.0           # JWT signing
httpx>=0.26.0                  # HTTP client (webhook transport)

# ============================================================
# Observability — new in v2.0
# ============================================================
opentelemetry-sdk>=1.20.0                          # core OTel
opentelemetry-api>=1.20.0
opentelemetry-exporter-otlp-proto-grpc>=1.20.0    # OTLP/gRPC exporter
opentelemetry-instrumentation-fastapi>=0.41b0      # auto-instrument FastAPI
opentelemetry-instrumentation-httpx>=0.41b0        # auto-instrument HTTP

# ============================================================
# AWS — new in v2.0 (production deployment)
# ============================================================
boto3>=1.34.0                  # AWS SDK (SNS, SQS, Bedrock, CloudWatch)
aws-xray-sdk>=2.12.0           # X-Ray distributed tracing (optional)

# ============================================================
# Utilities
# ============================================================
python-dotenv>=1.0.0

# ============================================================
# Streamlit Frontend
# ============================================================
streamlit>=1.40.0
streamlit-webrtc>=0.47.0
audio-recorder-streamlit>=0.0.10

# ============================================================
# Development & Testing
# ============================================================
pytest>=8.0.0
pytest-asyncio>=0.23.0         # async test support
pytest-cov>=4.1.0              # coverage reporting
httpx>=0.26.0                  # for FastAPI test client
black>=24.1.0
ruff>=0.1.14

# ============================================================
# Notebooks (for hands-on sessions)
# ============================================================
jupyter>=1.0.0
notebook>=7.0.0
ipywidgets>=8.0.0
matplotlib>=3.7.0
pandas>=2.0.0
